requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--extra-index-url https://alasdair.dalgarno%40babylonhealth.com:EofrkCaBbTvEoDG7XLKVZkoh@artifactory.ops.babylontech.co.uk/artifactory/api/pypi/babylon-pypi/simple

appdirs==1.4.4
    # via black
attrs==20.3.0
    # via
    #   -r requirements.in
    #   cattrs
    #   pytest
black==20.8b1
    # via -r requirements.in
cattrs==1.3.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   black
    #   flask
flask==1.1.2
    # via -r requirements.in
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.2
    # via -r requirements.in
regex==2020.11.13
    # via black
requests==2.25.1
    # via -r requirements.in
toml==0.10.2
    # via
    #   black
    #   pytest
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via black
urllib3==1.26.3
    # via requests
werkzeug==1.0.1
    # via flask
